requirements.txt
tensorflow~=2.3.1
opencv-python~=4.4.0.46
opencv-contrib-python
numpy~=1.18.5
six~=1.15.0
setuptools~=50.3.1
pillow~=8.0.1
matplotlib~=3.3.3
h5py~=2.10.0
contextlib2~=0.6.0.post1
ipython~=7.19.0
lxml~=4.6.1
dataclasses~=0.6
sentencepiece~=0.1.94
scipy~=1.5.4
pycocotools-windows~=2.0
pyyaml~=5.3.1
pandas~=1.1.4